requirements/base.txt
bcrypt==3.2.0
django-ninja==0.14.0
python-jose==3.3.0